Server-side batch rendering, test
