+++ src/Blazor.Extensions.Canvas/BECanvas.cshtml
@@ -1,3 +1,3 @@
 @inherits BECanvasComponent
 
-<canvas id="@Id" width="@Width" height="@Height" ref="canvasRef"></canvas>
+<canvas id="@Id" width="@Width" height="@Height" ref="_canvasRef"></canvas>
