+++ test/Blazor.Extensions.Canvas.Test.ClientSide/App.cshtml
