+++ test/Blazor.Extensions.Canvas.Test.ClientSide/Pages/Index.razor
@@ -1,0 +1,10 @@
+@page "/"
+@inherits IndexComponent
+
+<h1>Canvas demo!!!</h1>
+
+<a href="/webgl">View WebGL page</a><br>
+
+Welcome to your new app.
+
+<BECanvas Width="300" Height="400" ref="_canvasReference"></BECanvas>
