+++ test/Blazor.Extensions.Canvas.Test.ClientSide/Pages/IndexComponent.cs
@@ -1,0 +1,24 @@
+using Blazor.Extensions.Canvas.Canvas2D;
+using Microsoft.AspNetCore.Components;
+using System.Threading.Tasks;
+
+namespace Blazor.Extensions.Canvas.Test.ClientSide.Pages
+{
+    public class IndexComponent : ComponentBase
+    {
+        private Canvas2DContext _context;
+
+        protected BECanvasComponent _canvasReference;
+
+        protected override async Task OnAfterRenderAsync()
+        {
+            this._context = await this._canvasReference.CreateCanvas2DAsync();
+            await this._context.SetFillStyleAsync("green");
+
+            await this._context.FillRectAsync(10, 100, 100, 100);
+
+            await this._context.SetFontAsync("48px serif");
+            await this._context.StrokeTextAsync("Hello Blazor!!!", 10, 100);
+        }
+    }
+}
