+++ test/Blazor.Extensions.Canvas.Test.ClientSide/Pages/_ViewImports.cshtml
