+++ test/Blazor.Extensions.Canvas.Test.ClientSide/Shared/MainLayout.cshtml
@@ -1,0 +1,5 @@
+@inherits LayoutComponentBase
+
+<div>
+    @Body
+</div>
