+++ test/Blazor.Extensions.Canvas.Test.ClientSide/_ViewImports.cshtml
@@ -1,0 +1,9 @@
+@using System.Net.Http
+@using Microsoft.AspNetCore.Components.Forms
+@using Microsoft.AspNetCore.Components.Layouts
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.JSInterop
+@using Blazor.Extensions.Canvas.Test.ClientSide
+@using Blazor.Extensions.Canvas.Test.ClientSide.Shared
+@using Blazor.Extensions.Canvas
+@addTagHelper *, Blazor.Extensions.Canvas
