+++ test/Blazor.Extensions.Canvas.Test.ClientSide/wwwroot/index.html
@@ -1,4 +1,4 @@
-﻿<!DOCTYPE html>
+<!DOCTYPE html>
 <html>
 <head>
   <meta charset="utf-8" />
@@ -9,6 +9,6 @@
 <body>
   <app>Loading...</app>
 
-  <script src="_framework/blazor.webassembly.js"></script>
+  <script src="_framework/components.webassembly.js"></script>
 </body>
 </html>
