+++ test/Blazor.Extensions.Canvas.Test.ServerSide/Components/App.razor
@@ -1,0 +1,5 @@
+﻿@*
+    The Router component displays whichever component has a @page
+    directive matching the current URI.
+*@
+<Router AppAssembly="typeof(Startup).Assembly" />
