+++ test/Blazor.Extensions.Canvas.Test.ServerSide/Components/Pages/Index.razor
@@ -7,4 +7,4 @@
 
 Welcome to your new app.
 
-<BECanvas Width="300" Height="400" ref="@canvasReference"></BECanvas>
+<BECanvas Width="300" Height="400" ref="_canvasReference"></BECanvas>
