+++ test/Blazor.Extensions.Canvas.Test.ServerSide/Components/Pages/WebGL.razor
@@ -1,0 +1,10 @@
+@page "/webgl"
+@inherits WebGLComponent
+
+<h1>Canvas demo!!!</h1>
+
+<a href="/">View Canvas2d page</a><br>
+
+Welcome to your new app.
+
+<BECanvas Width="800" Height="600" ref="_canvasReference"></BECanvas>
