+++ test/Blazor.Extensions.Canvas.Test.ServerSide/Components/Pages/_ViewImports.cshtml
@@ -1,0 +1,1 @@
+@layout MainLayout
