+++ test/Blazor.Extensions.Canvas.Test.ServerSide/Pages/Index.cshtml
@@ -1,0 +1,16 @@
+@page "{*clientPath}"
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Blazor.Extensions.Canvas.Test.ServerSide</title>
+    <base href="~/" />
+    <script src="~/dist/blazor.extensions.canvas.js"></script>
+</head>
+<body>
+    <app>@(await Html.RenderComponentAsync<App>())</app>
+
+    <script src="_framework/components.server.js"></script>
+</body>
+</html>
