+++ test/Blazor.Extensions.Canvas.Test.ServerSide/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using Blazor.Extensions.Canvas.Test.ServerSide.Components
+@namespace Blazor.Extensions.Canvas.Test.ServerSide.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
