--- test/Blazor.Extensions.Canvas.Test/Pages/WebGL.cshtml
@@ -1,10 +1,0 @@
-@page "/webgl"
-@inherits WebGLComponent
-
-<h1>Canvas demo!!!</h1>
-
-<a href="/">View Canvas2d page</a><br>
-
-Welcome to your new app.
-
-<BECanvas Width="800" Height="600" ref="@canvasReference"></BECanvas>
