--- test/Blazor.Extensions.Canvas.Test/Shared/MainLayout.cshtml
@@ -1,5 +1,0 @@
-@inherits BlazorLayoutComponent
-
-<div>
-    @Body
-</div>
