--- test/Blazor.Extensions.Canvas.Test/_ViewImports.cshtml
@@ -1,9 +1,0 @@
-@using System.Net.Http
-@using Microsoft.AspNetCore.Blazor
-@using Microsoft.AspNetCore.Blazor.Components
-@using Microsoft.AspNetCore.Blazor.Layouts
-@using Microsoft.AspNetCore.Blazor.Routing
-@using Blazor.Extensions.Canvas.Test
-@using Blazor.Extensions.Canvas.Test.Shared
-@using Blazor.Extensions.Canvas
-@addTagHelper *, Blazor.Extensions.Canvas
